+++ script/proto/diagnostic.lua
@@ -78,6 +78,7 @@ m.register {
     'cast-type-mismatch',
     'return-type-mismatch',
     'inject-field',
+    'unnecessary-assert',
 } {
     group    = 'type-check',
     severity = 'Warning',