+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
+`2025-3-13`
+* `NEW` `unnecessary-assert` diagnostic warns when asserting values that are always truthy
 
 ## 3.13.9
 `2025-3-13`